rename verification method
